+++ Tyuiu.BrukhovAA.Sprint3.Task1.V15/Program.cs
@@ -30,7 +30,7 @@ namespace Tyuiu.BrukhovAA.Sprint3.Task1.V15
             Console.WriteLine("* ИСХОДНЫЕ ДАННЫЕ:                                                       *");
             Console.WriteLine("**************************************************************************");
 
-            double value = 2;
+            int value = 2;
             int startValue = 1;
             int stopValue = 5;
